backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
python-dotenv==1.0.1
pydantic==2.10.3
python-multipart==0.0.9
bcrypt==4.0.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
python-dotenv
gspread
pymongo[async]>=4.9,<4.10
motor
pandas==2.2.2